+++ pkg/codegen/internal/test/testdata/external-resource-schema/nodejs/cat.ts
@@ -53,9 +53,7 @@ export class Cat extends pulumi.CustomResource {
         } else {
             resourceInputs["name"] = undefined /*out*/;
         }
-        if (!opts.version) {
-            opts = pulumi.mergeOptions(opts, { version: utilities.getVersion()});
-        }
+        opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
         super(Cat.__pulumiType, name, resourceInputs, opts);
     }
 }