+++ pkg/codegen/internal/test/testdata/naming-collisions/nodejs/resource.ts
@@ -48,9 +48,7 @@ export class Resource extends pulumi.CustomResource {
         } else {
             resourceInputs["bar"] = undefined /*out*/;
         }
-        if (!opts.version) {
-            opts = pulumi.mergeOptions(opts, { version: utilities.getVersion()});
-        }
+        opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
         super(Resource.__pulumiType, name, resourceInputs, opts);
     }
 }