+++ pkg/codegen/internal/test/testdata/output-funcs-edgeorder/nodejs/listConfigurations.ts
@@ -14,9 +14,7 @@ export function listConfigurations(args: ListConfigurationsArgs, opts?: pulumi.I
         opts = {}
     }
 
-    if (!opts.version) {
-        opts.version = utilities.getVersion();
-    }
+    opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
     return pulumi.runtime.invoke("myedgeorder::listConfigurations", {
         "configurationFilters": args.configurationFilters,
         "customerSubscriptionDetails": args.customerSubscriptionDetails,