+++ pkg/codegen/internal/test/testdata/output-funcs/nodejs/listStorageAccountKeys.ts
@@ -14,9 +14,7 @@ export function listStorageAccountKeys(args: ListStorageAccountKeysArgs, opts?: 
         opts = {}
     }
 
-    if (!opts.version) {
-        opts.version = utilities.getVersion();
-    }
+    opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
     return pulumi.runtime.invoke("mypkg::listStorageAccountKeys", {
         "accountName": args.accountName,
         "expand": args.expand,