+++ pkg/codegen/internal/test/testdata/resource-args-python/nodejs/person.ts
@@ -52,9 +52,7 @@ export class Person extends pulumi.CustomResource {
             resourceInputs["name"] = undefined /*out*/;
             resourceInputs["pets"] = undefined /*out*/;
         }
-        if (!opts.version) {
-            opts = pulumi.mergeOptions(opts, { version: utilities.getVersion()});
-        }
+        opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);
         super(Person.__pulumiType, name, resourceInputs, opts);
     }
 }