+++ pkg/codegen/internal/test/testdata/simple-resource-schema/python/pulumi_example/foo_resource.py
@@ -74,6 +74,8 @@ class FooResource(pulumi.ComponentResource):
             raise TypeError('Expected resource options to be a ResourceOptions instance')
         if opts.version is None:
             opts.version = _utilities.get_version()
+        if opts.plugin_download_url is None:
+            opts.plugin_download_url = _utilities.get_plugin_download_url()
         if opts.id is not None:
             raise ValueError('ComponentResource classes do not support opts.id')
         else: