+++ pkg/codegen/nodejs/gen.go
@@ -865,9 +865,9 @@ func (mod *modContext) genResource(w io.Writer, r *schema.Resource) error {
 	fmt.Fprintf(w, "        }\n")
 
 	// If the caller didn't request a specific version, supply one using the version of this library.
-	fmt.Fprintf(w, "        if (!opts.version) {\n")
-	fmt.Fprintf(w, "            opts = pulumi.mergeOptions(opts, { version: utilities.getVersion()});\n")
-	fmt.Fprintf(w, "        }\n")
+	// If a `pluginDownloadURL` was supplied by the generating schema, we supply a default facility
+	// much like for version. Both operations are handled in the utilities library.
+	fmt.Fprint(w, "        opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);\n")
 
 	// Now invoke the super constructor with the type, name, and a property map.
 	if len(r.Aliases) > 0 {
@@ -1097,9 +1097,7 @@ func (mod *modContext) genFunction(w io.Writer, fun *schema.Function) error {
 	fmt.Fprintf(w, "        opts = {}\n")
 	fmt.Fprintf(w, "    }\n")
 	fmt.Fprintf(w, "\n")
-	fmt.Fprintf(w, "    if (!opts.version) {\n")
-	fmt.Fprintf(w, "        opts.version = utilities.getVersion();\n")
-	fmt.Fprintf(w, "    }\n")
+	fmt.Fprintf(w, "    opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts);\n")
 
 	// Now simply invoke the runtime function with the arguments, returning the results.
 	fmt.Fprintf(w, "    return pulumi.runtime.invoke(\"%s\", {\n", fun.Token)
@@ -1725,7 +1723,7 @@ func (mod *modContext) gen(fs fs) error {
 	case "":
 		buffer := &bytes.Buffer{}
 		mod.genHeader(buffer, nil, nil, nil)
-		fmt.Fprintf(buffer, "%s", utilitiesFile)
+		mod.genUtilitiesFile(buffer)
 		fs.add(path.Join(modDir, "utilities.ts"), buffer.Bytes())
 
 		// Ensure that the top-level (provider) module directory contains a README.md file.
@@ -2467,7 +2465,8 @@ func GeneratePackage(tool string, pkg *schema.Package, extraFiles map[string][]b
 	return files, nil
 }
 
-const utilitiesFile = `
+func (mod *modContext) genUtilitiesFile(w io.Writer) {
+	const body = `
 export function getEnv(...vars: string[]): string | undefined {
     for (const v of vars) {
         const value = process.env[v];
@@ -2513,4 +2512,16 @@ export function getVersion(): string {
     }
     return version;
 }
+
+/** @internal */
+export function resourceOptsDefaults(): any {
+    return { version: getVersion()%s };
+}
 `
+	var pluginDownloadURL string
+	if url := mod.pkg.PluginDownloadURL; url != "" {
+		pluginDownloadURL = fmt.Sprintf(", pluginDownloadURL: %q", url)
+	}
+	_, err := fmt.Fprintf(w, body, pluginDownloadURL)
+	contract.AssertNoError(err)
+}