+++ sdk/python/lib/pulumi/invoke.py
@@ -34,11 +34,17 @@ class InvokeOptions:
     An optional version. If provided, the provider plugin with exactly this version will be used to service
     the invocation.
     """
+    plugin_download_url : Optional[str]
+    """
+    An optional URL. If provided, the provider plugin with exactly this download URL will be used to service
+    the invocation. This will override the URL sourced from the host package, and should be rarely used.
+    """
 
     def __init__(self,
                  parent: Optional['Resource'] = None,
                  provider: Optional['ProviderResource'] = None,
-                 version: Optional[str] = "") -> None:
+                 version: Optional[str] = "",
+                 plugin_download_url: Optional[str] = None) -> None:
         """
         :param Optional[Resource] parent: An optional parent to use for default options for this invoke (e.g. the
                default provider to use).
@@ -46,7 +52,11 @@ class InvokeOptions:
                supplied, the default provider for the invoked function's package will be used.
         :param Optional[str] version: An optional version. If provided, the provider plugin with exactly this version
                will be used to service the invocation.
+        :param Optional[str] plugin_download_url: An optional URL. If provided, the provider plugin with this download
+               URL will be used to service the invocation. This will override the URL sourced from the host package, and
+               should be rarely used.
         """
         self.parent = parent
         self.provider = provider
         self.version = version
+        self.plugin_download_url = plugin_download_url